# keyaki/OpenAPI_Server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Keyaki Market - OpenAPI 3.0
  description: API documentation for PMUD final project Keyaki market.
  version: 1.0.11
servers:
- url: http://localhost:3000/
...
paths:
  /auth/user:
    get:
      tags:
      - auth
      summary: Get the logged-in user
      operationId: getUser
      responses:
        "200":
          description: Successfully retrieved user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "401":
          description: No user logged in
      x-swagger-router-controller: Auth
...
  /customers:
    get:
      tags:
      - customers
      summary: Get all customers
      operationId: getAllCustomers
      responses:
        "200":
          description: List of all customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200'
                x-content-type: application/json
        "500":
          description: Error retrieving customers
      x-swagger-router-controller: Customers
    post:
      tags:
      - customers
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        description: Customer data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/customers_body'
      responses:
        "201":
          description: Customer created successfully
        "400":
          description: Error creating customer
      x-swagger-router-controller: Customers
  /customers/{id}:
    get:
      tags:
      - customers
      summary: Get a customer by ID
      operationId: getCustomerById
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "404":
          description: Customer not found
        "500":
          description: Error retrieving customer
      x-swagger-router-controller: Customers
...
components:
  schemas:
    inline_response_200:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        password:
          type: string
      example:
        password: password
        name: name
        _id: _id
...
    customers_body:
      type: object
      properties:
        name:
          type: string
        password:
          type: string
...